virglrenderer: Avoid binary when libepoxy is +quartz
libvirglrenderer.1.dylib fails when it is installed from a binary
archive compiled on the buildserver against libepoxy +x11, but the user
has libepoxy +quartz installed:

| Termination Reason:    Namespace DYLD, Code 4 Symbol missing
| Symbol not found: _epoxy_glXChooseFBConfig
| Referenced from: <44D61F46-2561-3257-886E-71513200EEFC> /opt/local/lib/libvirglrenderer.1.dylib
| Expected in:     <789F52DB-263C-319B-B70E-560271F22A8F> /opt/local/lib/libepoxy.0.dylib

Rebuilding virglrenderer locally fixes this, but we should prevent the
use of the binary archive in this situation to avoid breakage on
a user's system.

Introduce empty +x11 and +quartz variants and automatically switch the
default variant based on the one chosen for libepoxy.