lima: Update to 1.1.0
#### Description

Upgrade to lima 1.1.0. This introduces `+additional_guestagents`, as
recommended in the ["Hint for package maintainers"][1] link, where
native guest agents are built and installed by default, but additional
architecture guest agents are not.

For macOS < 14, this variant is enabled by default; with macOS >= 14 it
is disabled by default. **This may be considered a breaking change.**

This is similar to the change where we disabled the installation of
`qemu` by default with Lima 1.0, but I am happy to change the
configuration so that the guest agents are built by default, but I would
also remove `depends_run-append port:qemu` from the variant (and add it
back to the macOS < 14 option).

If it's possible to have the build bots build both the default and
`+additional_guestagents` variants, this would make things much easier
for all users.

For what it's worth, adding `+additional_guestagents` more than triples
the build time (206s to 692s).

[1]: https://github.com/lima-vm/lima/releases/tag/v1.1.0#:~:text=Hint%20for%20package%20maintainers

###### Tested on

macOS 15.5 24F74 arm64
Xcode 16.3 16E140

###### Verification <!-- (delete not applicable items) -->

Have you

- [x] followed our [Commit Message Guidelines](https://trac.macports.org/wiki/CommitMessages)?
- [x] squashed and [minimized your commits](https://guide.macports.org/#project.github)?
- [x] checked that there aren't other open [pull requests](https://github.com/macports/macports-ports/pulls) for the same change?
- [x] referenced existing tickets on [Trac](https://trac.macports.org/wiki/Tickets) with full URL in commit message? <!-- Please don't open a new Trac ticket if you are submitting a pull request. -->
- [x] checked your Portfile with `port lint`?
- [x] tried a full install with `sudo port -vst install`?
- [x] tested basic functionality of all binary files?
- [x] checked that the Portfile's most important [variants](https://trac.macports.org/wiki/Variants) haven't been broken?