fix the macro so that 10.8 and higher call cloexec
seems to work by using an older-style.

also use the CloseallFDsOnExec that originated in chrome, which will
close all FDs after execution. though they say it's not safe in
multithreaded applications, it worked for quite some time and is better
than nothing.

all that's left is ensuring clang-19 uses the latest ld throughout.
seems to call ld127 at the end :/