borgbackup: Update to 1.4.1, use Python 3.13, ignore msgpack version
Update borgbackup to 1.4.1, which now supports Python 3.13, which is
currently our preferred version.

Enable running borgbackup tests, since I'm adding a patch to remove the
msgpack version check that keeps breaking otherwise working
installations and causing churn. See the patch header of the added
patchfile for an in-depth look at the situation to determine whether
this check is a net benefit for MacPorts.

Closes: https://trac.macports.org/ticket/72682