borgbackup: Update to 1.4.1, use Python 3.13, ignore msgpack version
Update borgbackup to 1.4.1, which now supports Python 3.13, which is
currently our preferred version.

Enable running borgbackup tests, since I'm adding a patch to remove the
msgpack version check that keeps breaking otherwise working
installations and causing churn. See the patch header of the added
patchfile for an in-depth look at the situation to determine whether
this check is a net benefit for MacPorts.

Change the reinplaces to fix the docs builds to use a patchfile, which
is cleaner and allows us to document the rationale in the patch file
header. Remove the reinplace that switches the sphinx-build executable
by overriding the Makefile variable on the command line instead to fix
an interrupted build, see https://trac.macports.org/ticket/72139.

Closes: https://trac.macports.org/ticket/72139
Closes: https://trac.macports.org/ticket/72682
Closes: https://trac.macports.org/ticket/72684