ensure emulated tls is built for 10.6
because of @jeremyhu's magic, or i think it's his magic, and a libc++
appearing in /usr/lib, the build process for clang-19 erroneously thinks
it can use it for cxa_thread_atexit, which it cannot. so i had to make
sure that we build clang to include it, as non-APPUL/CYGWIN platforms
use it, and snow leopard has no thread_local handling (that's on 10.7
and up).

so now we have a proper compiler that handles thread_local, and i'm
focussing on why environment.c in nodejs22 is not linking up to
libc++abi properly, which has the symbols.

all in all this is great news and it seems to me, at least, the compielr
is ready to go live. i just have to figure out how to make sure the
c++abi gets linked into environment.o before it gets into the static
library, which is then used at final linktime.

maybe jeremy has some ideas?

cc: rydye (@ryandesign)