add patches for remaining nodejses for 10.6, add @ryandesign's fixes
thanks pal.

proud to announce the new spiffy macports-libcxx holds up well to all of
the nodejses, unlike the existing one which cannot cope with the lack of
__cxa_thread_atexit, whereas mine, whilst not fully there (i would like
the symbol pulled automatically from the dylib instead of shoehorning in
the static), can pull the symbol out of the static libcxxabi.