lol @ macos14 failing on nodejs24 when 15 doesn't lol
there was "regression" in clang20 at some point, and i assume the
snapshot of APPUL clang contained it, as the other two don't fail on
nodejs24 (well, at least 15 doesn't on the last run).

so we force clang-19 on macos15 for nodejs24.

rootie @jmroot if this finishes when i'm in bed (and you'll obviously be up and
at it), wait to crack those Fosters kegs

cc @ryandesign