py-pyside6: don't create subports based on platform
Changing the version dynamically is bad enough, but changing which
subports exist makes it impossible to list all existing ports without
evaluating every Portfile on every possible platform. As a result,
things like the mirroring job fail if not run on a platform where the
subports are defined.

Instead, always create the subports and set platforms appropriately for
those that do not work on some platforms.