py-zipfile-deflate64: Xcode ≥ 16.3 compatibility
This adapts
https://github.com/madler/zlib/commit/4bd9a71f3539b5ce47f0c67ab5e01f3196dc8ef9
(2024-01-14), which appeared in zlib 1.3.1 (2024-01-22), to the snapshot of zlib
in zipfile-deflate64 0.2.0 (2022-01-05). This zlib snapshot is based on zlib
1.2.11 (2017-01-15).

This patch is necessary for compatibility with the Clang compiler in Xcode ≥
16.3: TARGET_OS_MAC as it was used in zutil.h intends to refer to the Classic
Mac OS. This macro has long been defined by the macOS SDK to mean modern macOS,
but was only defined by <TargetConditionals.h>, which was not ordinarily
https://github.com/llvm/llvm-project/commit/6e1f19168bca7e3bd4eefda50ba03eac8441dbbf
(2023-12-07, in upstream Clang 18.1, 2024-01-29, and in Xcode 16.3, 2025-03-31)
now has the compiler itself defining this macro unconditionally for Apple
targets
(https://releases.llvm.org/18.1.0/tools/clang/docs/ReleaseNotes.html#clang-frontend-potentially-breaking-changes),
with the effect of breaking code like this that used the macro in a different
way. There was some discussion of this at
https://github.com/madler/zlib/pull/895.