legacy-support-devel: Update to latest master - v20250924.
Notable changes since last -devel:

- Updates SDK version handling for SDK 26.

Due to the strictness of the version detection, this update is
required for successful builds with SDK 26.  But since
`legacy-support` doesn't offer any enhancements for OS versions later
than 10.15, this is only a problem for clients that unncesessarily use
it for later OS versions.

Also switches `github.tarball_from` to `archive` for the -devel
version, after verifying that this works correctly (earlier attempts
didn't).  The release version is left unchanged for now.

TESTED:
Tested both normal and -devel versions -/+universal on 10.4-10.5 ppc,
10.4-10.5 ppc64, 10.4-10.6 ppc (i386 Rosetta), 10.4-10.6 i386,
10.4-12.x x86_64, and 11.x-26.x arm64.
Builds and passes all tests on all tested platforms, except for the
aforementioned problem building the release version for OS 26.