borgbackup: Fix msgpack patch, add missing dependency
The msgpatch patch in its current form now always failed unless
BORG_MSGPACK_VERSION_CHECK=no was in the environment. Fix it to restore
the behavior of only checking for known broken versions (which currently
are none) or the environment variable.

Also, add a dependency on python-brotlicffi, since borgbackup doesn't
build without it in trace mode.

Fixes: https://github.com/macports/macports-ports/commit/4de29f91512e856b3ecc0931115342172f56b185